share code between food and textile
