+++ data/food/export.py
@@ -3,33 +3,30 @@
 
 """Ingredients and processes export for food"""
 
-import json
 import os
 import sys
-import urllib.parse
 from os.path import dirname
 
 import bw2calc
 import bw2data
-import matplotlib
-import numpy
-import pandas as pd
-import requests
 from bw2data.project import projects
 from common.export import (
+    IMPACTS_DEF_ECOBALYSE,
     cached_search,
+    check_ids,
+    compare_impacts,
+    compute_impacts,
+    csv_export_impact_comparison,
     display_changes,
     export_json_ordered,
+    find_id,
     load_json,
+    plot_impacts,
     progress_bar,
     remove_detailed_impacts,
-    spproject,
     with_aggregated_impacts,
     with_corrected_impacts,
-    with_subimpacts,
 )
-from common.impacts import bytrigram, main_method
-from common.impacts import impacts as definitions
 from frozendict import frozendict
 
 from food.ecosystemic_services.ecosystemic_services import (
@@ -51,8 +48,6 @@ if not ECOBALYSE_DATA_DIR:
 PROJECT = "default"
 DEFAULT_DB = "Agribalyse 3.1.1"
 ACTIVITIES_FILE = f"{PROJECT_ROOT_DIR}/data/food/activities.json"
-COMPARED_IMPACTS_FILE = f"{PROJECT_ROOT_DIR}/data/food/compared_impacts.csv"
-IMPACTS_FILE = f"{PROJECT_ROOT_DIR}/public/data/impacts.json"
 ECOSYSTEMIC_FACTORS_FILE = (
     f"{PROJECT_ROOT_DIR}/data/food/ecosystemic_services/ecosystemic_factors.csv"
 )
@@ -64,40 +59,19 @@ PROCESSES_AGGREGATED = f"{PROJECT_ROOT_DIR}/public/data/food/processes.json"
 LAND_OCCUPATION_METHOD = ("selected LCI results", "resource", "land occupation")
 GRAPH_FOLDER = f"{PROJECT_ROOT_DIR}/data/food/impact_comparison"
 
-with open(IMPACTS_FILE) as f:
-    IMPACTS_DEF_ECOBALYSE = json.load(f)
-
-
-def find_id(dbname, activity):
-    return cached_search(dbname, activity["search"]).get(
-        "Process identifier", activity["id"]
-    )
-
 
 def create_ingredient_list(activities_tuple):
     print("Creating ingredient list...")
     return tuple(
         [
-            process_activity_for_ingredient(activity)
+            to_ingredient(activity)
             for activity in list(activities_tuple)
             if activity["category"] == "ingredient"
         ]
     )
 
 
-def compute_normalization_factors():
-    normalization_factors = {}
-    for k, v in IMPACTS_DEF_ECOBALYSE.items():
-        if v["ecoscore"]:
-            normalization_factors[k] = (
-                v["ecoscore"]["weighting"] / v["ecoscore"]["normalization"]
-            )
-        else:
-            normalization_factors[k] = 0
-    return normalization_factors
-
-
-def process_activity_for_ingredient(activity):
+def to_ingredient(activity):
     return {
         "id": activity["id"],
         "name": activity["name"],
@@ -145,29 +119,12 @@ def compute_land_occupation(activities_tuple):
     return tuple(updated_activities)
 
 
-def check_ids(ingredients):
-    # Check the id is lowercase and does not contain space
-    for ingredient in ingredients:
-        if (
-            ingredient["id"].lower() != ingredient["id"]
-            or ingredient["id"].replace(" ", "") != ingredient["id"]
-        ):
-            raise ValueError(
-                f"This identifier is not lowercase or contains spaces: {ingredient['id']}"
-            )
-
-
 def create_process_list(activities):
     print("Creating process list...")
-    return frozendict(
-        {
-            activity["id"]: process_activity_for_processes(activity)
-            for activity in activities
-        }
-    )
+    return frozendict({activity["id"]: to_process(activity) for activity in activities})
 
 
-def process_activity_for_processes(activity):
+def to_process(activity):
     return {
         "id": activity["id"],
         "name": cached_search(activity.get("database", DEFAULT_DB), activity["search"])[
@@ -199,180 +156,6 @@ def process_activity_for_processes(activity):
     }
 
 
-def compute_simapro_impacts(activity, method):
-    strprocess = urllib.parse.quote(activity["name"], encoding=None, errors=None)
-    project = urllib.parse.quote(spproject(activity), encoding=None, errors=None)
-    method = urllib.parse.quote(main_method, encoding=None, errors=None)
-    return bytrigram(
-        definitions,
-        json.loads(
-            requests.get(
-                f"http://simapro.ecobalyse.fr:8000/impact?process={strprocess}&project={project}&method={method}"
-            ).content
-        ),
-    )
-
-
-def compute_brightway_impacts(activity, method):
-    results = dict()
-    lca = bw2calc.LCA({activity: 1})
-    lca.lci()
-    for key, method in definitions.items():
-        lca.switch_method(method)
-        lca.lcia()
-        results[key] = float("{:.10g}".format(lca.score))
-    return results
-
-
-def compare_impacts(processes_fd):
-    """This is compute_impacts slightly modified to store impacts from both bw and wp"""
-    processes = dict(processes_fd)
-    print("Computing impacts:")
-    for index, (key, process) in enumerate(processes.items()):
-        progress_bar(index, len(processes))
-        # simapro
-        activity = cached_search(process.get("source", DEFAULT_DB), process["search"])
-        results = compute_simapro_impacts(activity, main_method)
-        print(f"got impacts from SimaPro for: {process['name']}")
-        # WARNING assume remote is in m3 or MJ (couldn't find unit from COM intf)
-        if process["unit"] == "kilowatt hour" and isinstance(results, dict):
-            results = {k: v * 3.6 for k, v in results.items()}
-        if process["unit"] == "litre" and isinstance(results, dict):
-            results = {k: v / 1000 for k, v in results.items()}
-
-        process["simapro_impacts"] = results
-
-        # brightway
-        process["brightway_impacts"] = compute_brightway_impacts(activity, main_method)
-        print(f"got impacts from Brightway for: {process['name']}")
-
-        # compute subimpacts
-        process["simapro_impacts"] = with_subimpacts(process["simapro_impacts"])
-        process["brightway_impacts"] = with_subimpacts(process["brightway_impacts"])
-
-    processes_corrected_simapro = with_corrected_impacts(
-        IMPACTS_DEF_ECOBALYSE, processes, "simapro_impacts"
-    )
-    processes_corrected_smp_bw = with_corrected_impacts(
-        IMPACTS_DEF_ECOBALYSE, processes_corrected_simapro, "brightway_impacts"
-    )
-
-    return frozendict({k: frozendict(v) for k, v in processes_corrected_smp_bw.items()})
-
-
-def compute_impacts(processes_fd):
-    """Add impacts to processes dictionary
-
-    Args:
-        processes_fd (frozendict): dictionary of processes of which we want to compute the impacts
-    Returns:
-    dictionary of processes with impacts. Example :
-
-    {"sunflower-oil-organic": {
-        "id": "sunflower-oil-organic",
-        name": "...",
-        "impacts": {
-            "acd": 3.14,
-            ...
-            "ecs": 34.3,
-        },
-        "unit": ...
-        },
-    "tomato":{
-    ...
-    }
-    """
-    processes = dict(processes_fd)
-    print("Computing impacts:")
-    for index, (_, process) in enumerate(processes.items()):
-        progress_bar(index, len(processes))
-        # simapro
-        activity = cached_search(process.get("source", DEFAULT_DB), process["search"])
-        results = compute_simapro_impacts(activity, main_method)
-        # WARNING assume remote is in m3 or MJ (couldn't find unit from COM intf)
-        if process["unit"] == "kilowatt hour" and isinstance(results, dict):
-            results = {k: v * 3.6 for k, v in results.items()}
-        if process["unit"] == "litre" and isinstance(results, dict):
-            results = {k: v / 1000 for k, v in results.items()}
-
-        process["impacts"] = results
-
-        if isinstance(results, dict) and results:
-            # simapro succeeded
-            process["impacts"] = results
-            print(f"got impacts from simapro for: {process['name']}")
-        else:
-            # simapro failed (unexisting Ecobalyse project or some other reason)
-            # brightway
-            process["impacts"] = compute_brightway_impacts(activity, main_method)
-            print(f"got impacts from brightway for: {process['name']}")
-
-        # compute subimpacts
-        process["impacts"] = with_subimpacts(process["impacts"])
-
-        # remove unneeded attributes
-        for attribute in ["search"]:
-            if attribute in process:
-                del process[attribute]
-
-    return frozendict({k: frozendict(v) for k, v in processes.items()})
-
-
-def plot_impacts(ingredient_name, impacts_smp, impacts_bw):
-    impact_labels = impacts_smp.keys()
-    normalization_factors = compute_normalization_factors()
-
-    simapro_values = [
-        impacts_smp[label] * normalization_factors[label] for label in impact_labels
-    ]
-    brightway_values = [
-        impacts_bw[label] * normalization_factors[label] for label in impact_labels
-    ]
-
-    x = numpy.arange(len(impact_labels))
-    width = 0.35
-
-    fig, ax = matplotlib.pyplot.subplots(figsize=(12, 8))
-
-    ax.bar(x - width / 2, simapro_values, width, label="SimaPro")
-    ax.bar(x + width / 2, brightway_values, width, label="Brightway")
-
-    ax.set_xlabel("Impact Categories")
-    ax.set_ylabel("Impact Values")
-    ax.set_title(f"Environmental Impacts for {ingredient_name}")
-    ax.set_xticks(x)
-    ax.set_xticklabels(impact_labels, rotation=90)
-    ax.legend()
-
-    matplotlib.pyplot.tight_layout()
-    matplotlib.pyplot.savefig(f"{GRAPH_FOLDER}/{ingredient_name}.png")
-    matplotlib.pyplot.close()
-
-
-def csv_export_impact_comparison(compared_impacts):
-    rows = []
-    for product_id, process in compared_impacts.items():
-        simapro_impacts = process.get("simapro_impacts", {})
-        brightway_impacts = process.get("brightway_impacts", {})
-        for impact in simapro_impacts:
-            row = {
-                "id": product_id,
-                "name": process["name"],
-                "impact": impact,
-                "simapro": simapro_impacts.get(impact),
-                "brightway": brightway_impacts.get(impact),
-            }
-            row["diff_abs"] = abs(row["simapro"] - row["brightway"])
-            row["diff_rel"] = (
-                row["diff_abs"] / abs(row["simapro"]) if row["simapro"] != 0 else None
-            )
-
-            rows.append(row)
-
-    df = pd.DataFrame(rows)
-    df.to_csv(COMPARED_IMPACTS_FILE, index=False)
-
-
 if __name__ == "__main__":
     projects.set_current(PROJECT)
     bw2data.config.p["biosphere_database"] = "biosphere3"
@@ -399,16 +182,18 @@ if __name__ == "__main__":
     processes = create_process_list(activities_land_occ)
 
     if len(sys.argv) == 1:  # just export.py
-        processes_impacts = compute_impacts(processes)
+        processes_impacts = compute_impacts(processes, DEFAULT_DB)
     elif len(sys.argv) > 1 and sys.argv[1] == "compare":  # export.py compare
-        impacts_compared_dic = compare_impacts(processes)
-        csv_export_impact_comparison(impacts_compared_dic)
+        impacts_compared_dic = compare_impacts(processes, DEFAULT_DB)
+        csv_export_impact_comparison(impacts_compared_dic, "food")
         for ingredient_name, values in impacts_compared_dic.items():
             print(f"Plotting {ingredient_name}")
             simapro_impacts = values["simapro_impacts"]
             brightway_impacts = values["brightway_impacts"]
             os.makedirs(GRAPH_FOLDER, exist_ok=True)
-            plot_impacts(ingredient_name, simapro_impacts, brightway_impacts)
+            plot_impacts(
+                ingredient_name, simapro_impacts, brightway_impacts, GRAPH_FOLDER
+            )
             print("Charts have been generated and saved as PNG files.")
         sys.exit(0)
     else:
